dnsbl/bypass fix

+++ helpers/checks/dnsbl.js
@@ -13,13 +13,6 @@ module.exports = async (req, res, next) => {
 
 	if (dnsbl.enabled && dnsbl.blacklists.length > 0 //if dnsbl enabled and has more than 0 blacklists
 		&& !res.locals.anonymizer) { //anonymizers cant be dnsbl'd
-		if (blockBypass.bypassDnsbl) {
-			if (!res.locals.blockBypass) {
-				return blockBypassCheck(req, res, next);
-			}
-			return next(); //already solved
-		}
-		//otherwise dnsbl cant be bypassed
 		const ip = req.headers[ipHeader] || req.connection.remoteAddress;
 		let isBlacklisted = await cache.get(`blacklisted:${ip}`);
 		if (isBlacklisted === null) { //not cached
@@ -28,6 +21,13 @@ module.exports = async (req, res, next) => {
 			await cache.set(`blacklisted:${ip}`, isBlacklisted, Math.floor(dnsbl.cacheTime/1000));
 		}
 		if (isBlacklisted) {
+			if (blockBypass.bypassDnsbl) {
+				if (!res.locals.blockBypass) {
+					return blockBypassCheck(req, res, next);
+				}
+				return next(); //already solved
+			}
+			//otherwise dnsbl cant be bypassed
 			deleteTempFiles(req).catch(e => console.error);
 			return dynamicResponse(req, res, 403, 'message', {
 				'title': 'Forbidden',
